Styles change and bug fixes
- H1 margin style removal
- Change in pixel ratios in example data
- Crop buttons moved in a separated file
- Image preview container pixel ratios are larger
- New component added header title
- Input label style change
- Bug fixed drop outside component didn't activate on second dragged and dropped image
- Section overlay added for crop preview
