+++ src/components/data/exampleData.ts
@@ -1,10 +1,10 @@
 const exampleData: { title: string; value: string }[] = [
     { title: "Porcelain shop - 670x670px", value: "670x670" },
     { title: "Facebook cover - 812x312px", value: "812x312" },
-    { title: "Facebook ad - 1200x720px", value: "1200x720" },
+    { title: "Facebook ad - 720x720px", value: "720x720" },
     { title: "Instagram profile - 110x110px", value: "110x110" },
     { title: "Instagram post - 320x320px", value: "320x320" },
-    { title: "Instagram story - 1080x1920px", value: "1080x1920" },
+    { title: "Instagram story - 1000x1000px", value: "1000x1000" },
     { title: "X profile - 400x400px", value: "400x400" },
 ];
