+++ src/components/elements/HeaderTitle/HeaderTitle.tsx
@@ -1,0 +1,9 @@
+import style from "./HeaderTitle.module.css";
+
+export default function HeadarTitle() {
+    return (
+        <nav className={style.headarTitleContainer}>
+            <h1>Crop tool</h1>
+        </nav>
+    );
+}
